Implemented basic recipe and ingredient logic
I have created the basic logic for the recipes.
This involved the creation of the recipe model, controller, and migration.
All the recipes have a many to many relationship with ingredients.
Therefore, I have also created the model and migrations for the ingredients.
I have created a pivot table between the recipe and ingredient models.
This also contains the amount of each ingredient in each recipe.
I have created one route for the recipe, the store method.
This method has a corresponding feature test to ensure it works as intended.